Resolve "Using items"

Closes #23

See merge request efrei-2a-sdd-p2022-m0n/project!13

+++ src/inventory.h
@@ -3,10 +3,10 @@
 
 //Statistics, also used for the player
 typedef struct stats_entity {
-    int health;
-    int damage;
-    int speed;
-    int agility;
+    float health;
+    float damage;
+    float speed;
+    float agility;
 } stats_entity;
 
 //Functions relative to the inventory and the shop
